# openapi.yaml
openapi: 3.1.0
info:
  title: Incident Tracker API
  version: 0.3.0
  description: >-
    Multi-tenant incident tracker API with session auth, tenancy, and incident workflows.
    ID-based routes return 404 for objects that are missing or inaccessible in the active
    tenant context, and return 403 only when object existence is established but permission
    is denied.
servers:
  - url: http://localhost:3000
paths:
...
  /v1/incidents:
    get:
      summary: List incidents for active tenant
      operationId: listIncidents
      security:
        - SessionCookie: []
        - ApiKeyHeader: []
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Incident page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentListResponse'
        '400':
          description: Invalid tenant context or pagination cursor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Active tenant membership not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create incident
      operationId: createIncident
      security:
        - SessionCookie: []
          CsrfToken: []
        - ApiKeyHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncidentRequest'
      responses:
        '201':
          description: Incident created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentItemResponse'
        '400':
          description: Invalid tenant context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Tenant write quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Active tenant membership not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/incidents/{incidentId}:
    get:
      summary: Get incident by ID
      operationId: getIncident
      security:
        - SessionCookie: []
        - ApiKeyHeader: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: Incident details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentItemResponse'
        '403':
          description: Permission denied after incident lookup.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Incident not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      summary: Update incident fields
      operationId: updateIncident
      security:
        - SessionCookie: []
          CsrfToken: []
        - ApiKeyHeader: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIncidentRequest'
      responses:
        '200':
          description: Incident updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentItemResponse'
        '400':
          description: Invalid update payload or transition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Tenant write quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Incident not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/incidents/{incidentId}/timeline-events:
    get:
      summary: List incident timeline events
      operationId: listIncidentTimelineEvents
      security:
        - SessionCookie: []
        - ApiKeyHeader: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: Timeline events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEventsResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Incident not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Add timeline event
      operationId: createIncidentTimelineEvent
      security:
        - SessionCookie: []
          CsrfToken: []
        - ApiKeyHeader: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimelineEventRequest'
      responses:
        '201':
          description: Timeline event created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineEventItemResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Tenant write quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Incident not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/incidents/{incidentId}/tasks:
    get:
      summary: List incident tasks
      operationId: listIncidentTasks
      security:
        - SessionCookie: []
        - ApiKeyHeader: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: Incident tasks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentTasksResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Incident not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create incident task
      operationId: createIncidentTask
      security:
        - SessionCookie: []
          CsrfToken: []
        - ApiKeyHeader: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncidentTaskRequest'
      responses:
        '201':
          description: Task created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentTaskItemResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Tenant write quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Incident not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/incidents/{incidentId}/tasks/{taskId}:
    patch:
      summary: Update incident task
      operationId: updateIncidentTask
      security:
        - SessionCookie: []
          CsrfToken: []
        - ApiKeyHeader: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIncidentTaskRequest'
      responses:
        '200':
          description: Task updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentTaskItemResponse'
        '403':
          description: Permission denied after task lookup.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Tenant write quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Incident or task not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/incidents/{incidentId}/status-updates:
    get:
      summary: List incident status updates
      operationId: listIncidentStatusUpdates
      security:
        - SessionCookie: []
        - ApiKeyHeader: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      responses:
        '200':
          description: Status updates.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentStatusUpdatesResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Incident not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      summary: Create incident status update
      operationId: createIncidentStatusUpdate
      security:
        - SessionCookie: []
          CsrfToken: []
        - ApiKeyHeader: []
      parameters:
        - $ref: '#/components/parameters/IncidentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStatusUpdateRequest'
      responses:
        '201':
          description: Status update created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncidentStatusUpdateItemResponse'
        '403':
          description: Permission denied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Tenant write quota exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Incident not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
...
    IncidentId:
      name: incidentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: incident_tracker_sid
    CsrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
    ApiKeyHeader:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    CreateIncidentRequest:
      type: object
      required: [title, severity, startTime, impactedServices]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        description:
          type: string
          maxLength: 4000
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        startTime:
          type: string
          format: date-time
        impactedServices:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 100
    UpdateIncidentRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        description:
          type: string
          maxLength: 4000
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        endTime:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        impactedServices:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 100
    CreateTimelineEventRequest:
      type: object
      required: [eventTime, eventType, message]
      properties:
        eventTime:
          type: string
          format: date-time
        eventType:
          type: string
          minLength: 2
          maxLength: 100
        message:
          type: string
          minLength: 1
          maxLength: 4000
    CreateIncidentTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        description:
          type: string
          maxLength: 4000
        assigneeUserId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        dueAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
    UpdateIncidentTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        description:
          type: string
          maxLength: 4000
        status:
          $ref: '#/components/schemas/TaskStatus'
        assigneeUserId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        dueAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
    CreateStatusUpdateRequest:
      type: object
      required: [audience, message, publishedAt]
      properties:
        audience:
          $ref: '#/components/schemas/StatusUpdateAudience'
        message:
          type: string
          minLength: 1
          maxLength: 4000
        publishedAt:
          type: string
          format: date-time
...
    IncidentSeverity:
      type: string
      enum: [SEV1, SEV2, SEV3, SEV4]
    IncidentStatus:
      type: string
      enum: [declared, investigating, mitigating, monitoring, resolved, closed]
    TaskStatus:
      type: string
      enum: [open, in_progress, completed]
    StatusUpdateAudience:
      type: string
      enum: [internal, external]
...
    Incident:
      type: object
      required:
        [
          id,
          tenantId,
          title,
          description,
          severity,
          status,
          startTime,
          endTime,
          declaredByUserId,
          impactedServices,
          createdAt,
          updatedAt
        ]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        severity:
          $ref: '#/components/schemas/IncidentSeverity'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        startTime:
          type: string
          format: date-time
        endTime:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        declaredByUserId:
          type: string
          format: uuid
        impactedServices:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TimelineEvent:
      type: object
      required: [id, tenantId, incidentId, eventTime, eventType, message, createdByUserId, createdAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        incidentId:
          type: string
          format: uuid
        eventTime:
          type: string
          format: date-time
        eventType:
          type: string
        message:
          type: string
        createdByUserId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    IncidentTask:
      type: object
      required:
        [id, tenantId, incidentId, title, description, status, assigneeUserId, dueAt, createdByUserId, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        incidentId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        assigneeUserId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
        dueAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
        createdByUserId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    IncidentListResponse:
      type: object
      required: [incidents, nextCursor, csrfToken]
      properties:
        incidents:
          type: array
          items:
            $ref: '#/components/schemas/Incident'
        nextCursor:
          oneOf:
            - type: string
            - type: 'null'
        csrfToken:
          type: string
    IncidentItemResponse:
      type: object
      required: [incident, csrfToken]
      properties:
        incident:
          $ref: '#/components/schemas/Incident'
        csrfToken:
          type: string
    TimelineEventsResponse:
      type: object
      required: [events, csrfToken]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/TimelineEvent'
        csrfToken:
          type: string
    TimelineEventItemResponse:
      type: object
      required: [event, csrfToken]
      properties:
        event:
          $ref: '#/components/schemas/TimelineEvent'
        csrfToken:
          type: string
    IncidentTasksResponse:
      type: object
      required: [tasks, csrfToken]
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/IncidentTask'
        csrfToken:
          type: string
    IncidentTaskItemResponse:
      type: object
      required: [task, csrfToken]
      properties:
        task:
          $ref: '#/components/schemas/IncidentTask'
        csrfToken:
          type: string
    IncidentStatusUpdatesResponse:
      type: object
      required: [updates, csrfToken]
      properties:
        updates:
          type: array
          items:
            $ref: '#/components/schemas/IncidentStatusUpdate'
        csrfToken:
          type: string
    IncidentStatusUpdateItemResponse:
      type: object
      required: [statusUpdate, csrfToken]
      properties:
        statusUpdate:
          $ref: '#/components/schemas/IncidentStatusUpdate'
        csrfToken:
          type: string
...
    ErrorResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
        details:
          description: Optional structured metadata for debugging and validation.